refactor(vm-driver): the fake's staging area becomes a type of its own
The fake staged through `PreparedFake`'s own methods, which ties the
area to the grip that made it. Derive it from the runtime dir instead, so
a second grip on the same VM computes the same area rather than drifting
into a second one. Pure move: no behaviour changes.
