+++ virt/arcbox-vm-driver/src/testkit/mod.rs
@@ -19,6 +19,7 @@ pub mod contract;
 pub mod fake_driver;
 pub mod fake_network;
 mod fake_prepared;
+mod fake_staging;
 mod fake_vm;
 mod fake_vsock;
